`Renderer` module code organization
